functions for index.php file too
I was making fun changes and found if you remove the html designations, you can put an index.php in the hash folder instead (on a php server env of course [apache]). Don't know how secure this is though and I'm sure there is probably a better way to secure a php page. More research is needed. XD